Se reorganizan archivos, se añade la pantalla de AM inmueble (incompleta)
