+++ src/main/java/GUI/Panels/PantallaABMCliente.java
@@ -1,7 +1,7 @@
 package GUI.Panels;
 
-import GUI.GestorGUI;
-import GUI.Pantalla;
+import Services.GestorGUI;
+import GUI.Util.Pantalla;
 
 import javax.swing.*;
 import javax.swing.event.*;
@@ -9,8 +9,6 @@ import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableRowSorter;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 
 public class PantallaABMCliente {
