+++ src/main/java/GUI/Panels/PantallaAMInmueble.java
@@ -1,0 +1,13 @@
+package GUI.Panels;
+
+import javax.swing.*;
+
+public class PantallaAMInmueble {
+    private JPanel panelPrincipal;
+    private JPanel panelTitulos;
+    private JLabel titulo;
+
+    public JPanel getPanelPrincipal() {
+        return panelPrincipal;
+    }
+}
